chore: Moved `MultiMapToHighNeeds` function to its own `Mapper` class
